requirements.txt
config==0.5.0.post0
greenlet==1.0.0
numpy==1.20.2
pandas==1.2.4
path==15.1.2
path.py==12.5.0
pathlib==1.0.1
pysqlite3==0.4.6
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.4.9